Established OTM relation between databases, refactored feature spec helpers into a combined file.
