+++ app.rb
@@ -17,6 +17,7 @@ class MakersBNB < Sinatra::Base
     redirect '/' if params[:password] != params[:password_confirmation]
     User.create(email: params[:email], password: params[:password])
     session[:user] = User.authenticate(email: params[:email], password: params[:password])
+    p session[:user]
     redirect '/spaces'
   end
