+++ spec/features/signup_spec.rb
@@ -1,10 +1,10 @@
+require 'spec_helper'
+
+require_relative '../helpers/helpers'
+
 feature 'Signup' do
   scenario 'a new user signs up and is taken to /spaces page' do
-    visit('/')
-    fill_in 'email', with: 'user@makersbnb.com'
-    fill_in 'password', with: 'passw0rd'
-    fill_in 'password_confirmation', with: 'passw0rd'
-    click_button 'Sign up'
+    sign_up
     expect(page).to have_content("Book a Space")
   end
