+++ spec/features/space_booking_spec.rb
@@ -1,9 +1,11 @@
 require 'spec_helper'
-require_relative '../helpers/list_space'
+
+require_relative '../helpers/helpers'
 
 feature 'Booking a space' do
-  
+
   scenario 'the space user books a space' do
+    sign_up
     list_space
     click_link('spaceID_1')
     expect(page).to have_content 'Please book your space'
